docs(codespec): 彻底重写文档架构，53 份收敛为 14 份
重构 codespec 目录架构、消除重复描述、统一序号体系，并逐条核对文档与代码运行时的一致性。

架构重构：
- 53 份 .md 收敛为 14 份手写 + 1 份生成物（ERROR_CATALOG.md）
- 顶层 6 份均改为自包含正文，不再是外链空壳索引
- specification/ 取消 features/subsystems/subsystems_api 三分法，改为按 include/aurora/ 模块域平面罗列 8 份（01-core … 08-tooling）
- COMMIT_CONVENTION.md 并入 CODING_STANDARDS.md §10

消除重复：
- 合并 subsystems/ 与 subsystems_api/ 的 5 组同名分叉文件
- 合并 LAYOUT_PROTOCOL + SUBSYSTEM_LAYOUT + SUBSYSTEM_API_LAYOUT_ENGINE 三重布局文档
- features/（A–G）需求内容折入对应子系统文档后删除目录

序号体系：
- 章节号统一为纯数字点分层级（1 / 1.1 / 1.1.1），禁止中英文序号
- 保留 #1–#24 需求编号与验收标准，与章节号并存
- 全文交叉引用按新结构重建，零死链

代码与文档一致性（修正旧文档误写）：
- Result 无 is_ok()，只有 ok()；Result<void> 无 value()/unwrap()
- WidgetRegistry / JsonPatchOp 在 aurora::serialization 子命名空间
- patch path 含 /props/ 段；diff 返回裸 JSON 数组，无信封
- LengthKind 为四个无参枚举值，尺寸数值在 Length::value
- to_code 属 serialization 命名空间；Diagnostics 只有 report/warn/degraded
- 各控件提供静态 defaults()，Widget::defaults() 不存在

去除任务性内容：开发路线图、P0–P2 优先级、不可分割组、历史迁移注记、状态性与远期项措辞。

同步更新 AGENTS.md 的文档导航表、模块划分与硬规则的文档映射。

验证：#1–#24 齐备，零死链，cmake --build build --target gen_api_tools 通过。

+++ include/aurora/window/window.h
@@ -720,7 +720,7 @@ class Window {
             // 不应改变其尺寸；若变化说明它实际依赖子节点，误判为 boundary。
             const Size after = w.size();
             const bool size_unchanged =
-                std::fabsf(before.width - after.width) < 1e-3f && std::fabsf(before.height - after.height) < 1e-3f;
+                std::fabs(before.width - after.width) < 1e-3f && std::fabs(before.height - after.height) < 1e-3f;
             AURORA_ASSERT(
                 size_unchanged,
                 "relayout boundary 误判：boundary 尺寸随重排变化，说明其依赖子节点，不能作为 relayout boundary");
